refactor: create new files for each model
