+++ app/models/__init__.py
@@ -1,3 +1,5 @@
-from app.models.trips import Trips, TripDays, Activities, Base
+from app.models.trips import Trips, Base
+from app.models.activities import Activities
+from app.models.trip_days import TripDays
 from app.models.user import User
 from app.models.vote import Vote
